+++ packages/circuit-ui/components/Modal/Modal.mdx
@@ -106,6 +106,8 @@ Use the `immersive` variant to focus the user's attention on the dialog content.
 
 ## Related components
 
+This component is built on top of the low level [Dialog](Components/Dialog/Docs) component. If this component does not meet your requirements, you can use the `Dialog` component directly to build your own custom dialog.
+
 For cases displaying simple text content but requiring immediate or critical action(s) from the user, consider using the [NotificationModal](Notification/NotificationModal/Docs) component.
 
 ## Accessibility
